Improve the hashhing code.

+++ linera-views/src/map_view.rs
@@ -22,6 +22,7 @@ pub struct MapView<C, I, V> {
     was_cleared: bool,
     updates: BTreeMap<Vec<u8>, Option<V>>,
     _phantom: PhantomData<I>,
+    stored_hash: Option<HashOutput>,
     hash: Option<HashOutput>,
 }
 
@@ -45,6 +46,7 @@ where
             was_cleared: false,
             updates: BTreeMap::new(),
             _phantom: PhantomData,
+            stored_hash: hash,
             hash,
         })
     }
@@ -52,7 +54,7 @@ where
     fn rollback(&mut self) {
         self.was_cleared = false;
         self.updates.clear();
-        self.hash = None;
+        self.hash = self.stored_hash;
     }
 
     fn flush(&mut self, batch: &mut Batch) -> Result<(), ViewError> {
@@ -74,10 +76,13 @@ where
                 }
             }
         }
-        let key = self.context.base_tag(KeyTag::Hash as u8);
-        match self.hash {
-            None => batch.delete_key(key),
-            Some(hash) => batch.put_key_value(key, &hash)?,
+        if self.stored_hash != self.hash {
+            let key = self.context.base_tag(KeyTag::Hash as u8);
+            match self.hash {
+                None => batch.delete_key(key),
+                Some(hash) => batch.put_key_value(key, &hash)?,
+            }
+            self.stored_hash = self.hash;
         }
         Ok(())
     }
